Add files via upload
The python code you provided is a simple menu system for a Kota cafe. It imports four modules: chai, coffee, snacks, and coldd. Each module contains a function that returns a list of items available in that category.

The main function starts by printing a welcome message and a menu. Then, it enters a loop where it prompts the user to select an option. If the user selects a valid option, the program calls the corresponding function from the imported modules to print a list of items available in that category. The user can then select an item from the list.

If the user selects an invalid option, the program prints an error message. The loop continues until the user enters "n" to indicate that they do not want to continue.

Here is a more detailed description of each function:

chai(): This function returns a list of chai flavors available at the Kota cafe.
coffee(): This function returns a list of coffee flavors available at the Kota cafe
snacks(): This function returns a list of snacks available at the Kota cafe
cold(): This function returns a list of cold drinks available at the Kota cafe.
The main function also has a special case for the option 4, which is for water bottles. For this option, the program prompts the user to select the size of water bottle they want. Then, it prints the total price of the water bottle.

Overall, the code you provided is a simple but effective menu system for a Kota cafe It is easy to understand and modify, and it can be easily extended to include additional categories and items.